backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.8.0
pydantic_core==2.20.0
python-multipart==0.0.6
python-dotenv==1.0.0

motor==3.3.2
pymongo==4.5.0

pytest==7.4.3
pytest-asyncio==0.21.1

httpx==0.25.2
requests==2.31.0
